+++ App/views/alumni.py
@@ -1,76 +1,72 @@
 import os
-from flask import Blueprint, app, current_app, redirect, render_template, request, send_from_directory, jsonify, url_for, flash
+from flask import Blueprint, current_app, flash,  jsonify, redirect, render_template, request, url_for
 from App.models import db
 from werkzeug.utils import secure_filename
 
-from flask_jwt_extended import jwt_required, current_user, unset_jwt_cookies, set_access_cookies
+from flask_jwt_extended import current_user, jwt_required
 
-from App.models.notification import Notification
-
-from .index import index_views
-
-
-from App.controllers import(
-    get_user_by_email,
-    # update_alumni_info,
-    get_job_listing,
-    get_all_job_listings,
-    get_approved_listings,
-)
-
-from App.controllers.company_subscription import (
-    add_company_subscription,
-    get_company_subscription,
-    get_all_company_subscriptions,
-    get_company_subscriptions_by_alumnus_id,
-    delete_company_subscription
-)
-
-from App.controllers.saved_job_listing import get_saved_job_listings_by_alumnus_id
-from App.models import(
+from App.models import (
     AlumnusAccount,
     CompanyAccount,
-    AdminAccount,
+    Notification,
     SavedJobListing,
-    JobApplication,
-    CompanySubscription
+    JobApplication
 )
 
-alumni_views = Blueprint('alumni_views', __name__, template_folder='../templates')
+from App.controllers.alumnus_account import update_alumnus_info
+from App.controllers.base_user_account import get_user_by_email
+from App.controllers.company_subscription import (
+    add_company_subscription,
+    get_company_subscription
+)
+from App.controllers.job_listing import get_job_listing
+from App.controllers.saved_job_listing import get_saved_job_listings_by_alumnus_id
 
-# @alumni_views.route('/update_alumni/<id>', methods=['POST'])
-# @jwt_required()
-# def update_alumni(id):
-#     user = get_user_by_email(current_user.login_email)
-#     data = request.form
-#     first_name = data['fname']
-#     last_name = data['lname']
-#     phone_number = data['contact']
-#     login_email = data['email']
 
-#     current_password = data['current_password']
-#     confirm_current_password = data['confirm_current_password']
-#     new_password = data['new_password']
-#     confirm_new_password = data['confirm_new_password']
+alumni_views = Blueprint(
+    'alumni_views',
+    __name__,
+    template_folder='../templates'
+)
 
-#     if current_password != confirm_current_password:
-#         flash ("Current passwords do not match")
-#         return render_template('my-account-alumni.html', user=user)
 
-#     if new_password != confirm_new_password:
-#         flash ("New passwords do not match")
-#         return render_template('my-account-alumni.html', user=user)
+@alumni_views.route('/update_alumni/<id>', methods=['POST'])
+@jwt_required()
+def update_alumni(id):
+    if not isinstance(current_user, AlumnusAccount):
+        flash('Unauthorized access', 'unsuccessful')
+        return redirect(url_for('index_views.index_page'))
+    
+    user = get_user_by_email(current_user.login_email)
+    data = request.form
+    first_name = data['fname']
+    last_name = data['lname']
+    phone_number = data['contact']
+    login_email = data['email']
 
-#     update_status = update_alumni_info(id, first_name, last_name, phone_number, login_email, current_password, new_password)
+    current_password = data['current_password']
+    confirm_current_password = data['confirm_current_password']
+    new_password = data['new_password']
+    confirm_new_password = data['confirm_new_password']
 
-#     if update_status:
-#         flash ("Alumni information updated successfully")
-#         return render_template('my-account-alumni.html', user=user)
-#     else:
-#         flash("Update failed. Check your information and try again.")
-#         return render_template('my-account-alumni.html', user=user)
+    if current_password != confirm_current_password:
+        flash("Current passwords do not match")
+        return render_template('my-account-alumni.html', user=user)
 
-#     return render_template('my-account-alumni.html', user=user)
+    if new_password != confirm_new_password:
+        flash("New passwords do not match")
+        return render_template('my-account-alumni.html', user=user)
+
+    update_status = update_alumnus_info(
+        id, first_name, last_name, phone_number, login_email, current_password, new_password)
+
+    if update_status:
+        flash("Alumni information updated successfully")
+    else:
+        flash("Update failed. Check your information and try again.")
+
+    return render_template('my-account-alumni.html', user=user)
+
 
 @alumni_views.route('/view_my_account/<id>', methods=["GET"])
 @jwt_required()
@@ -81,9 +77,8 @@ def view_my_account_page(id):
 
     except Exception:
         flash('Error retreiving User')
-        response = redirect(url_for('index_views.index_page'))
+        return redirect(url_for('index_views.index_page'))
 
-    return response
 
 @alumni_views.route('/subscribe', methods=['POST'])
 @jwt_required()
@@ -91,47 +86,48 @@ def subscribe_action():
     """
     Allows an alumnus to subscribe to a company to receive updates about job listings.
     """
-     # Get the list of selected companies from the form
+    if not isinstance(current_user, AlumnusAccount):
+        flash('Unauthorized access', 'unsuccessful')
+        return redirect(url_for('index_views.index_page'))
+    
     selected_companies = request.form.getlist('company')
-    # print(f"Selected companies: {selected_companies}")  # Test print
-    
-    
+
     if not selected_companies:
         flash('No companies selected!', 'error')
     try:
         # Get the alumnus ID and company ID from the request
         alumnus_id = current_user.id
-        
+
         # Check if the alumnus and company exist in the database
         alumnus = AlumnusAccount.query.get(alumnus_id)
-        companies = CompanyAccount.query.filter(CompanyAccount.registered_name.in_(selected_companies)).all()
-        # print(f"Queried companies: {companies}")  # Test print
-        
+        companies = CompanyAccount.query.filter(
+            CompanyAccount.registered_name.in_(selected_companies)).all()
+
         if not alumnus:
-           flash('Alumnus not found!', 'error') 
-           return redirect(url_for('index_views.index_page'))
-        
+            flash('Alumnus not found!', 'error')
+            return redirect(url_for('index_views.index_page'))
+
         if not companies:
             flash('No valid companies selected!', 'error')
             return redirect(url_for('index_views.index_page'))
 
         new_subscription_added = False
-         
+
         for company in companies:
             # Check if the alumnus is already subscribed to the company
-            existing_subscription = get_company_subscription(alumnus_id=alumnus.id, company_id=company.id)
-        
+            existing_subscription = get_company_subscription(
+                alumnus_id=alumnus.id, company_id=company.id)
+
             if existing_subscription:
                 continue  # Skip if the alumnus is already subscribed to this company
 
             # Create a new subscription
-            subscription = add_company_subscription(alumnus_id=alumnus.id, company_id=company.id)
+            add_company_subscription(
+                alumnus_id=alumnus.id, company_id=company.id)
             new_subscription_added = True  # Mark that a new subscription has been added
-        
+
         db.session.commit()
-        
-        # subscriptions = CompanySubscription.query.all()
-        # print(subscriptions)  # Test print
+
         if new_subscription_added:
             flash('Successfully subscribed to selected companies!', 'success')
         else:
@@ -144,47 +140,27 @@ def subscribe_action():
         flash(f"An error occurred: {str(e)}", 'unsuccessful')
         return redirect(url_for('index_views.index_page'))
 
-# @alumni_views.route('/unsubscribe', methods=['POST'])
-# @jwt_required()
-# def unsubscribe_action():
-#     # get form data
-#     # data = request.form
-#     response = None
-
-#     # print(data)
-
-#     try:
-#         alumni = delete_company_subscription(current_user.id)
-#         # print(alumni.get_json())
-#         response = redirect(url_for('index_views.index_page'))
-#         flash('Unsubscribed!', 'success')
-
-#     except Exception:
-#         # db.session.rollback()
-#         flash('Error unsubscribing', 'unsuccessful')
-#         response = redirect(url_for('auth_views.login_page'))
-
-#     return response
 
 # @alumni_views.route('/update_modal_seen', methods=['POST'])
 # @jwt_required()
 # def update_modal_seen():
 #     try:
-#         alumni = current_user  
-#         set_alumni_modal_seen(alumni.id)  
-#         db.session.commit() 
+#         alumni = current_user
+#         set_alumni_modal_seen(alumni.id)
+#         db.session.commit()
 #         return jsonify(message="Modal seen status updated successfully"), 200
-        
+
 #     except Exception as e:
-#         db.session.rollback()  
+#         db.session.rollback()
 #         return jsonify(message="Error updating modal status"), 500
-    
+
+
 @alumni_views.route('/view_listing_alumni/<id>', methods=["GET"])
 @jwt_required()
 def view_listing_page(id):
-    listing=get_job_listing(id)
-    saved_listings=get_saved_job_listings_by_alumnus_id(current_user.id)
-    
+    listing = get_job_listing(id)
+    saved_listings = get_saved_job_listings_by_alumnus_id(current_user.id)
+
     try:
         return render_template('view-listing-alumni.html', listing=listing, saved_listings=saved_listings)
 
@@ -198,44 +174,63 @@ def view_listing_page(id):
 @alumni_views.route('/get_saved_listing', methods=['GET'])
 @jwt_required()
 def get_saved_job_listing():
-    alumnus_id = current_user.id
+    if not isinstance(current_user, AlumnusAccount):
+        flash('Unauthorized access', 'unsuccessful')
+        return redirect(url_for('index_views.index_page'))
 
-    already_saved = SavedJobListing.query.filter_by(alumnus_id=alumnus_id).all()
+    already_saved = SavedJobListing.query.filter_by(
+        alumnus_id=current_user.id).all()
 
     return jsonify([listing.job_listing_id for listing in already_saved])
+
 
 @alumni_views.route('/save_listing/<job_listing_id>', methods=['POST'])
 @jwt_required()
 def save_job_listing(job_listing_id):
+    if not isinstance(current_user, AlumnusAccount):
+        flash('Unauthorized access', 'unsuccessful')
+        return redirect(url_for('index_views.index_page'))
+    
     alumnus_id = current_user.id
 
-    already_saved = SavedJobListing.query.filter_by(alumnus_id=alumnus_id, job_listing_id=job_listing_id).first()
+    already_saved = SavedJobListing.query.filter_by(
+        alumnus_id=alumnus_id, job_listing_id=job_listing_id).first()
 
     if not already_saved:
-        new_saved_job_listing = SavedJobListing(alumnus_id=alumnus_id, job_listing_id=job_listing_id)
+        new_saved_job_listing = SavedJobListing(
+            alumnus_id=alumnus_id, job_listing_id=job_listing_id)
         db.session.add(new_saved_job_listing)
         db.session.commit()
 
         return jsonify({"message": "Job saved successfully!", "status": "saved"}), 201
 
+
 @alumni_views.route('/remove_saved_listing/<job_listing_id>', methods=['GET'])
 @jwt_required()
 def remove_listing(job_listing_id):
-    alumnus_id = current_user.id
+    if not isinstance(current_user, AlumnusAccount):
+        flash('Unauthorized access', 'unsuccessful')
+        return redirect(url_for('index_views.index_page'))
 
-    already_saved_job = SavedJobListing.query.filter_by(alumnus_id=alumnus_id, job_listing_id=job_listing_id).first()
+    already_saved_job = SavedJobListing.query.filter_by(
+        alumnus_id=current_user.id, job_listing_id=job_listing_id).first()
 
     if not already_saved_job:
         return jsonify({"message": "Job not saved!", "status": "error"}), 404
-    
+
     db.session.delete(already_saved_job)
     db.session.commit()
 
-    return jsonify({"message": "Job Removed from saved listings", "status":"removed"}), 200
+    return jsonify({"message": "Job Removed from saved listings", "status": "removed"}), 200
+
 
 @alumni_views.route('/apply_to_listing/<int:job_listing_id>', methods=['POST'])
 @jwt_required()
 def apply(job_listing_id):
+    if not isinstance(current_user, AlumnusAccount):
+        flash('Unauthorized access', 'unsuccessful')
+        return redirect(url_for('index_views.index_page'))
+    
     # Get form data
     work_experience = request.form.get("work-experience")
     resume = request.files["resume"]
@@ -244,7 +239,7 @@ def apply(job_listing_id):
     filename = secure_filename(resume.filename)
     static_folder = os.path.join(current_app.root_path, 'static')
     resume_path = os.path.join(static_folder, 'uploads', 'resumes', filename)
-    file_path =  os.path.join('uploads', 'resumes', filename)
+    file_path = os.path.join('uploads', 'resumes', filename)
 
     # Save the file to the directory
     resume.save(resume_path)  # This actually writes the file!
@@ -256,7 +251,7 @@ def apply(job_listing_id):
     new_application = JobApplication(
         alumnus_id=alumnus_id,
         job_listing_id=job_listing_id,
-        resume_file_path=file_path,  
+        resume_file_path=file_path,
         work_experience=work_experience,
     )
 
@@ -269,34 +264,35 @@ def apply(job_listing_id):
     flash("Application submitted successfully!", "success")
     return redirect(url_for("index_views.index_page"))
 
+
 @alumni_views.route('/alumnus_notifications', methods=['GET'])
 @jwt_required()
 def view_notifications_page():
-    
-    alumnus = current_user
-    
-    if not isinstance(alumnus,AlumnusAccount):
-        flash('Not an Alumnus', 'unsuccessful')
+    if not isinstance(current_user, AlumnusAccount):
+        flash('Unauthorized access', 'unsuccessful')
         return redirect(url_for('index_views.index_page'))
 
     try:
         # Fetch notifications for the alumnus
-        notifications = alumnus.notifications.all()
+        notifications = current_user.notifications.all()
         return render_template('alumnus_notifications.html', notifications=notifications, alumnus=current_user)
-    
+
     except Exception as e:
         flash('Error retrieving notifications', 'unsuccessful')
         return redirect(url_for('index_views.index_page'))
 
+
 @alumni_views.route('/check_unread_notifications', methods=['GET'])
 @jwt_required()
 def check_notifications():
+    if not isinstance(current_user, AlumnusAccount):
+        flash('Unauthorized access', 'unsuccessful')
+        return redirect(url_for('index_views.index_page'))
+    
     # Fetch unread notifications for the current user
-    unread_notifications = Notification.query.filter_by(alumnus_id=current_user.id, reviewed_by_user=False).all()
+    unread_notifications = Notification.query.filter_by(
+        alumnus_id=current_user.id, reviewed_by_user=False).all()
 
     # Determine if there are new notifications
     has_new_notifications = len(unread_notifications) > 0
     return jsonify({'has_new_notifications': has_new_notifications})
-    
-
-
